+++ BLCMM/src/CHANGELOG.md
@@ -59,6 +59,8 @@ BLCMM Changelog
   shown on the active screen (this prevents clearing the recently-viewed file
   list, for instance)
 - Removed hidden ability to save in legacy FilterTool mod format
+- Changed behavior when encountering invalid strings in map/level merge
+  statements -- should possibly be more accepting now.
 
 (changelog entries to confirm, etc -- from original 1.2.1 plans ages ago)
 - When creating and saving a new file, it now opens on the next boot